# open-api/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/asset/upload": {
      "post": {
        "operationId": "uploadFile",
        "parameters": [
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetDto"
              }
            }
          },
          "description": "Asset Upload Information",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetFileUploadResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.94.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "AssetFileUploadResponseDto": {
        "properties": {
          "duplicate": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "duplicate",
          "id"
        ],
        "type": "object"
      },
...
      "CreateAssetDto": {
        "properties": {
          "assetData": {
            "format": "binary",
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isExternal": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isOffline": {
            "type": "boolean"
          },
          "isReadOnly": {
            "type": "boolean"
          },
          "isVisible": {
            "type": "boolean"
          },
          "libraryId": {
            "format": "uuid",
            "type": "string"
          },
          "livePhotoData": {
            "format": "binary",
            "type": "string"
          },
          "sidecarData": {
            "format": "binary",
            "type": "string"
          }
        },
        "required": [
          "assetData",
          "deviceAssetId",
          "deviceId",
          "fileCreatedAt",
          "fileModifiedAt"
        ],
        "type": "object"
      },
...
    }
  }
}